+++ src/components/UI/BaseInput/BaseInput.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { fn } from '@storybook/test';
-import { BaseInput } from '../components/UI/BaseInput/BaseInput';
+import { BaseInput } from './BaseInput';
 
 const meta: Meta<typeof BaseInput> = {
   component: BaseInput,
