+++ src/components/UI/Button/Button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { fn } from '@storybook/test';
-import { Button } from '../components/UI/Button/Button';
+import { Button } from './Button';
 
 const meta: Meta<typeof Button> = {
   component: Button,
